requirements.txt
   python-dotenv
   together